# apps/server/src/index.ts
// ============== SENTRY (must be first) ==============
import * as Sentry from '@sentry/node';

import express from 'express';
import cors from 'cors';
import cookieParser from 'cookie-parser';
import helmet from 'helmet';
import dotenv from 'dotenv';
import { createServer } from 'http';
import { Server } from 'socket.io';
import rateLimit from 'express-rate-limit';

// Load env vars FIRST
dotenv.config();

// ============== ENV VALIDATION (fail-fast) ==============
const REQUIRED_ENVS = [
    'DATABASE_URL',
    'JWT_ACCESS_SECRET',
    'JWT_REFRESH_SECRET',
];

const RECOMMENDED_ENVS = [
    'FRONTEND_URL',
    'RTMP_SECRET',
];

for (const envName of REQUIRED_ENVS) {
    if (!process.env[envName]) {
        console.error(`\n❌ FATAL: Missing required environment variable: ${envName}`);
        console.error(`   Please set it in your .env file or environment.\n`);
        process.exit(1);
    }
}

for (const envName of RECOMMENDED_ENVS) {
    if (!process.env[envName]) {
        console.warn(`⚠️  WARNING: Missing recommended env var: ${envName} (using fallback)`);
    }
}

// Import configs
import prisma from './config/prisma';
import { connectRedis } from './config/redis';

// Import routes
import authRoutes from './routes/authRoutes';
import userRoutes from './routes/userRoutes';
import donationRoutes from './routes/donationRoutes';
import walletRoutes from './routes/walletRoutes';
import streamRoutes from './routes/streamRoutes';
import adminRoutes from './admin/adminRoutes';
import webhookRoutes from './payments/webhookRoutes';

const app = express();
const httpServer = createServer(app);

// ============== SENTRY INIT ==============
if (process.env.SENTRY_DSN) {
    Sentry.init({
        dsn: process.env.SENTRY_DSN,
        environment: process.env.NODE_ENV || 'development',
        tracesSampleRate: process.env.NODE_ENV === 'production' ? 0.2 : 1.0,
        release: `kwanza-stream@${process.env.npm_package_version || '1.0.0'}`,
    });
    console.log('✅ Sentry initialized');
} else {
    console.warn('⚠️  SENTRY_DSN not set — error tracking disabled');
}

// Socket.io setup
const io = new Server(httpServer, {
    cors: {
        origin: process.env.FRONTEND_URL || 'http://localhost:3000',
        methods: ['GET', 'POST'],
        credentials: true,
    },
});

// ============== SECURITY MIDDLEWARE ==============
app.use(helmet({
    contentSecurityPolicy: {
        directives: {
            defaultSrc: ["'self'"],
            scriptSrc: ["'self'"],
            styleSrc: ["'self'", "'unsafe-inline'"], // Required for inline styles in UI components
            imgSrc: ["'self'", "data:", "blob:", "https:"],
            connectSrc: ["'self'", process.env.FRONTEND_URL || 'http://localhost:3000'],
            fontSrc: ["'self'", "https://fonts.gstatic.com"],
            mediaSrc: ["'self'", "blob:"],
            objectSrc: ["'none'"],
            frameAncestors: ["'none'"],
            upgradeInsecureRequests: process.env.NODE_ENV === 'production' ? [] : null,
        },
    },
    crossOriginEmbedderPolicy: false, // Required for cross-origin media loading
}));
app.use(cors({
    origin: process.env.FRONTEND_URL || 'http://localhost:3000',
    credentials: true,
}));
app.use(express.json({ limit: '1mb' }));
app.use(cookieParser());

// XSS protection — strip HTML from all inputs
import { sanitizeMiddleware } from './middleware/sanitizeMiddleware';
app.use(sanitizeMiddleware);

// ============== RATE LIMITING ==============
// Auth endpoints: strict limit to prevent OTP/login brute-force
const authLimiter = rateLimit({
    windowMs: 15 * 60 * 1000, // 15 minutes
    max: 10,                   // 10 attempts per 15 min
    message: { error: 'Muitas tentativas. Aguarda 15 minutos.' },
    standardHeaders: true,
    legacyHeaders: false,
    keyGenerator: (req) => {
        // Combine IP + userId (if present) for shared networks
        const ip = req.ip || req.socket.remoteAddress || 'unknown';
        const userId = (req.body as any)?.phone || (req.body as any)?.identifier || '';
        return `${ip}:${userId}`;
    },
});

// Moderate limit for financial endpoints (wallet, donations)
const financialLimiter = rateLimit({
    windowMs: 1 * 60 * 1000, // 1 minute
    max: 30,                  // 30 requests per minute
    message: { error: 'Too many financial requests. Please slow down.' },
    standardHeaders: true,
    legacyHeaders: false,
});

// Reports: anti-spam
const reportLimiter = rateLimit({
    windowMs: 1 * 60 * 1000,
    max: 5,
    message: { error: 'Muitos relatórios. Aguarda 1 minuto.' },
    standardHeaders: true,
    legacyHeaders: false,
});

// Search: anti-scraping
const searchLimiter = rateLimit({
    windowMs: 1 * 60 * 1000,
    max: 30,
    message: { error: 'Muitas pesquisas. Aguarda um momento.' },
    standardHeaders: true,
    legacyHeaders: false,
});

// Stream creation
const streamCreateLimiter = rateLimit({
    windowMs: 1 * 60 * 1000,
    max: 20,
    message: { error: 'Too many stream requests.' },
    standardHeaders: true,
    legacyHeaders: false,
});

// General API limit — 100 req/min per IP
const generalLimiter = rateLimit({
    windowMs: 1 * 60 * 1000,
    max: 100,
    message: { error: 'Demasiados pedidos. Tenta novamente em breve.' },
    standardHeaders: true,
    legacyHeaders: false,
});

// Apply general limiter to all API routes
app.use('/api/', generalLimiter);

// ============== HEALTH CHECKS ==============
app.get('/', (req, res) => {
    res.json({
        message: 'Kwanza Stream API is running 🚀',
        version: '1.0.0',
        timestamp: new Date().toISOString(),
    });
});

// Detailed health endpoint
import healthRoutes from './routes/healthRoutes';
app.use('/api/health', healthRoutes);

// ============== API ROUTES ==============
app.use('/api/auth', authLimiter, authRoutes);
app.use('/api/users', userRoutes);
app.use('/api/donations', financialLimiter, donationRoutes);
app.use('/api/wallet', financialLimiter, walletRoutes);
app.use('/api/streams', streamCreateLimiter, streamRoutes);
app.use('/api/admin', adminRoutes);
app.use('/api/webhooks', webhookRoutes); // Webhooks have their own HMAC auth

// Phase 1: Upload routes + static file serving
import uploadRoutes from './routes/uploadRoutes';
app.use('/api/upload', uploadRoutes);
app.use('/uploads', express.static('uploads'));

// Phase 2: Notification + Favorite routes
import notificationRoutes from './routes/notificationRoutes';
import favoriteRoutes from './routes/favoriteRoutes';
app.use('/api/notifications', notificationRoutes);
app.use('/api/favorites', favoriteRoutes);

// Phase 5: Report routes
import reportRoutes from './routes/reportRoutes';
app.use('/api/reports', reportLimiter, reportRoutes);

// Phase 7: Search routes
import searchRoutes from './routes/searchRoutes';
app.use('/api/search', searchLimiter, searchRoutes);

// P1: Creator Studio + Clips
import creatorRoutes from './routes/creatorRoutes';
import clipsRoutes from './routes/clipsRoutes';
app.use('/api/creator', creatorRoutes);
app.use('/api/clips', clipsRoutes);

// P2: Events
import eventsRoutes from './routes/eventsRoutes';
app.use('/api/events', eventsRoutes);

// Sentry test endpoint (dev/staging only)
if (process.env.NODE_ENV !== 'production') {
    app.get('/api/debug-sentry', (_req, _res) => {
        throw new Error('Sentry test error — if you see this in Sentry, it works! 🎉');
    });
}

// Import streaming services
import { startMediaServer } from './streaming/mediaServer';
import { setupChatService } from './streaming/chatService';
import { registerModerationHandlers } from './streaming/chatModerationController';

// Setup chat service
setupChatService(io);

// P0: Register chat moderation handlers for each new connection
io.on('connection', (socket) => {
    registerModerationHandlers(io, socket);
});

// Sentry error handler — must be before custom error handler
if (process.env.SENTRY_DSN) {
    Sentry.setupExpressErrorHandler(app);
}

// Error handling middleware
app.use((err: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
    console.error('Server error:', err);
    res.status(500).json({ error: 'Internal server error' });
});

// Start server
const PORT = process.env.PORT || 5000;

const startServer = async () => {
    try {
        // Connect to Redis (optional, continues if fails)
        try {
            await connectRedis();
        } catch (err) {
            console.warn('Redis connection failed, continuing without Redis');
        }

        // Verify database connection
        await prisma.$connect();
        console.log('✅ Database connected');

        // Start RTMP media server (optional, may fail if ffmpeg not installed)
        try {
            startMediaServer();
        } catch (err) {
            console.warn('⚠️  RTMP server failed to start (FFmpeg may not be installed)');
        }

        httpServer.listen(PORT, () => {
            console.log(`
🚀 Kwanza Stream API Server
📡 Port: ${PORT}
🌍 Environment: ${process.env.NODE_ENV || 'development'}
📊 Health: http://localhost:${PORT}/health
🔐 Auth: http://localhost:${PORT}/api/auth (rate limited: 10/15min)
💰 Wallet: http://localhost:${PORT}/api/wallet (rate limited: 30/min)
🎬 Streams: http://localhost:${PORT}/api/streams
      `);
        });
    } catch (error) {
        console.error('Failed to start server:', error);
        process.exit(1);
    }
};

startServer();

// ============== GRACEFUL SHUTDOWN ==============
const shutdown = async (signal: string) => {
    console.log(`\n${signal} received — shutting down gracefully...`);
    httpServer.close(() => console.log('✅ HTTP server closed'));
    await prisma.$disconnect();
    console.log('✅ Database disconnected');
    process.exit(0);
};

process.on('SIGTERM', () => shutdown('SIGTERM'));
process.on('SIGINT', () => shutdown('SIGINT'));
process.on('unhandledRejection', (reason) => {
    console.error('Unhandled Rejection:', reason);
    if (process.env.SENTRY_DSN) Sentry.captureException(reason);
});
process.on('uncaughtException', (error) => {
    console.error('Uncaught Exception:', error);
    if (process.env.SENTRY_DSN) Sentry.captureException(error);
    process.exit(1);
});

export { io };