# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kwanza Stream API
  description: |
    API da plataforma Kwanza Stream — streaming social angolano.
    Autenticação via httpOnly cookies (access_token + refresh_token).
  version: 1.0.0
  contact:
    name: Kwanza Stream Team
servers:
  - url: http://localhost:5000
    description: Development
  - url: https://staging-api.kwanzastream.com
    description: Staging
  - url: https://api.kwanzastream.com
    description: Production
...
paths:
...
  /api/donations:
    post:
      tags: [Donations]
      summary: Enviar Salo (doação)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiverId, saloType]
              properties:
                receiverId: { type: string, format: uuid }
                saloType: { type: string, enum: [bronze, silver, gold, diamond, legendary] }
                message: { type: string, maxLength: 200 }
                streamId: { type: string, format: uuid }
      responses:
        '200':
          description: Salo enviado com sucesso
...
  /api/streams:
    post:
      tags: [Streams]
      summary: Criar nova stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                category: { type: string }
      responses:
        '201':
          description: Stream criada

  # ============== NOTIFICATIONS ==============
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Listar notificações (cursor pagination)
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lista de notificações
...
  /api/search:
    get:
      tags: [Search]
      summary: Pesquisar utilizadores e streams
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 2 }
        - in: query
          name: type
          schema: { type: string, enum: [all, users, streams], default: all }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: Resultados de pesquisa

  # ============== ADMIN ==============
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Dashboard statistics
      responses:
        '200':
          description: Estatísticas gerais
        '403':
          description: Não autorizado (requer ADMIN role)
  /api/admin/transactions/{id}/approve:
    post:
      tags: [Admin]
      summary: Aprovar levantamento pendente
      description: Requer permissão admin.transactions.approve (FINANCIAL_ADMIN ou SUPER_ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Levantamento aprovado
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set by /api/auth/login or /api/auth/verify-otp
security:
  - cookieAuth: []